add probeManager

+++ @types/editor-extends.d.ts
@@ -51,7 +51,7 @@ interface EditorExtendsComponent {
 interface EditorExtendsAsset {
     queryAssetInfo(uuid: string, callback: Function): any;
     getAssetInfoFromUrl(url: string): EditorAssetInfo;
-    saveDataToImage(buffer: Float32Array | null, width: number, height: number, fullpath: string, callback: Function): any;
+    saveDataToImage(buffer: Uint8Array | null, width: number, height: number, fullpath: string, callback: Function): any;
 }
 
 interface EditorExtendsUuid {
